requirements.txt
# 多智能体框架依赖包
torch>=1.9.0
transformers>=4.20.0
huggingface-hub>=0.10.0
datasets>=2.0.0
accelerate>=0.12.0
pyyaml>=6.0
asyncio-mqtt>=0.11.0
pytest>=7.0.0
pytest-asyncio>=0.20.0